Make DynamodB TTL relative to event timestamp
Until now, the reference timestamp used to compute the TTL was an
absolute one generated in the dispatcher after the output of Kinesis.
Consequently, in the worst case, the real TTL of the mapping between the
IID and the OID could be more than 48h (24h of Kinesis + 24h15min of
DynamoDB).
This modification uses the relative timestamp of the ingested event as
the reference for calculating the TTL and fixes the aforementioned
issue.

Thanks to @latk for reporting the issue.

Signed-off-by: Farzad Senart <515714+fsenart@users.noreply.github.com>